About Me & Resume Page
Successfully built the core functionality and basic look of the "About Me" webpage for the portfolio.

Successfully built the core functionality and basic look of the "Resume" webpage for the portfolio.

The "Resume" webpage uses an iframe to embed a .pdf viewer as well as a download button for the resume .pdf file itself.

Created a .CSS file for each webpage respectively. Each file has the same name as the .html file it corresponds to.

Added some miscellaneous media files  such as a filler image for the About Me profile picture and my resume .pdf file for embedding.