Webpage styling
Added a darker shade of white for the webpage backgrounds.

Added some padding below the back buttons on the Resume and Contact webpages.

Added a black background color to the webpages that aren't scrollable.

Removed the extra height added by the header css tag so it now matches up with the height of the original nav bar container.